refactor: migrate backend packages to project root
Move file_management/ and networks/ packages and node.py from
p2p_ui/base/backend/ to the project root alongside main.py.

- Update node.py imports to absolute (networks.peer, file_management.*)
- Update views.py to add project root to sys.path and import Node from root
- Update tests.py imports and patch paths to match new module locations
- Replace old CLI prototype main.py with the current headless runner
- Remove empty p2p_ui/base/backend/ directory
